Don't try to pass size_t to printf.
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@14804 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/sc_player.cpp
@@ -6622,8 +6622,8 @@ bool player_t::parse_talents_wowhead( const std::string& talent_string )
 
     if ( ( total < '0' ) || ( total > '/' + 3 * ( 1 + 4 + 16) ) )
     {
-      sim -> errorf( "Player %s has malformed wowhead talent string '%s': encoded character '%c' in position %ld is invalid.\n",
-                     name(), talent_string.c_str(), total, idx );
+      sim -> errorf( "Player %s has malformed wowhead talent string '%s': encoded character '%c' in position %d is invalid.\n",
+                     name(), talent_string.c_str(), total, ( int )idx );
       return false;
     }
